Test Functions Added + Error Solved. Constraint #4?
The 2 test functions have been added and the assertions are working.

Singular Matrix C in LSQ Subproblem is also no longer returned when the
function is called.

The program seems to be working at this point. One interesting thing is
that when constraint 4 is removed from the ieqcons when the fmin_slsqp
optimization method, the value of the minimum changes greatly and the
maximum of the function is of the same magnitude as the minimum. There
are also too many iterations to terminate the optimization
successfully.When constraint 4 is removed, the program seems to work a
lot more smoothly.